refactor: split trade desk into audience and realtime conversions destinations (#3091)
* refactor: split trade desk into audience and realtime conversions destinations

* fix: add headers for real time conversions

* chore: apply review suggestions
