+++ src/cdk/v2/destinations/the_trade_desk/transformRecord.js
@@ -1,4 +1,4 @@
-const { InstrumentationError, ConfigurationError } = require('@rudderstack/integrations-lib');
+const { InstrumentationError } = require('@rudderstack/integrations-lib');
 const { BatchUtils } = require('@rudderstack/workflow-engine');
 const {
   defaultPostRequestConfig,
@@ -12,20 +12,6 @@ const { getTTLInMin, getFirstPartyEndpoint } = require('./utils');
 const tradeDeskConfig = require('./config');
 
 const { DATA_PROVIDER_ID } = tradeDeskConfig;
-
-const validateConfig = (config) => {
-  if (!config.advertiserSecretKey) {
-    throw new ConfigurationError('Advertiser Secret Key is not present. Aborting');
-  }
-
-  if (config.ttlInDays && !(config.ttlInDays >= 0 && config.ttlInDays <= 180)) {
-    throw new ConfigurationError('TTL is out of range. Allowed values are 0 to 180 days');
-  }
-
-  if (!config.audienceId) {
-    throw new ConfigurationError('Segment name/Audience ID is not present. Aborting');
-  }
-};
 
 const responseBuilder = (items, config) => {
   const { advertiserId, dataServer } = config;
@@ -63,8 +49,6 @@ const processRecordInputs = (inputs, destination) => {
   if (!inputs || inputs.length === 0) {
     return [];
   }
-
-  validateConfig(Config);
 
   const invalidActionTypeError = new InstrumentationError(
     'Invalid action type. You can only add or remove IDs from the audience/segment',
