+++ src/cdk/v2/destinations/the_trade_desk_real_time_conversions/data/TTDItemConfig.json
