+++ src/v0/destinations/the_trade_desk/networkHandler.js
@@ -8,36 +8,27 @@ const { getSignatureHeader } = require('../../../cdk/v2/destinations/the_trade_d
 const { isHttpStatusSuccess } = require('../../util/index');
 const tags = require('../../util/tags');
 const { JSON_MIME_TYPE } = require('../../util/constant');
-const {
-  REAL_TIME_CONVERSION_ENDPOINT,
-} = require('../../../cdk/v2/destinations/the_trade_desk/config');
 
 const proxyRequest = async (request) => {
   const { endpoint, data, method, params, headers, config } = prepareProxyRequest(request);
-  let ProxyHeaders = {
+
+  if (!config?.advertiserSecretKey) {
+    throw new PlatformError('Advertiser secret key is missing in destination config. Aborting');
+  }
+
+  if (!process.env.THE_TRADE_DESK_DATA_PROVIDER_SECRET_KEY) {
+    throw new PlatformError('Data provider secret key is missing. Aborting');
+  }
+
+  const ProxyHeaders = {
     ...headers,
     'Content-Type': JSON_MIME_TYPE,
+    TtdSignature: getSignatureHeader(data, config.advertiserSecretKey),
+    'TtdSignature-dp': getSignatureHeader(
+      data,
+      process.env.THE_TRADE_DESK_DATA_PROVIDER_SECRET_KEY,
+    ),
   };
-
-  // For first party data flow
-  if (endpoint !== REAL_TIME_CONVERSION_ENDPOINT) {
-    if (!config?.advertiserSecretKey) {
-      throw new PlatformError('Advertiser secret key is missing in destination config. Aborting');
-    }
-
-    if (!process.env.THE_TRADE_DESK_DATA_PROVIDER_SECRET_KEY) {
-      throw new PlatformError('Data provider secret key is missing. Aborting');
-    }
-
-    ProxyHeaders = {
-      ...ProxyHeaders,
-      TtdSignature: getSignatureHeader(data, config.advertiserSecretKey),
-      'TtdSignature-dp': getSignatureHeader(
-        data,
-        process.env.THE_TRADE_DESK_DATA_PROVIDER_SECRET_KEY,
-      ),
-    };
-  }
 
   const requestOptions = {
     url: endpoint,
@@ -69,7 +60,6 @@ const responseHandler = (destinationResponse) => {
   // Trade desk first party data api returns 200 with an error in case of "Failed to parse TDID, DAID, UID2, IDL, EUID, or failed to decrypt UID2Token or EUIDToken"
   // https://partner.thetradedesk.com/v3/portal/data/doc/post-data-advertiser-external
   // {"FailedLines":[{"ErrorCode":"MissingUserId","Message":"Invalid DAID, item #1"}]}
-  // For real time conversion api we don't have separate response handling, trade desk always return 400 for bad events.
   if ('FailedLines' in response && response.FailedLines.length > 0) {
     throw new AbortedError(
       `Request failed with status: ${status} due to ${JSON.stringify(response)}`,
