+++ test/integrations/destinations/the_trade_desk/common.ts
@@ -3,7 +3,6 @@ const destTypeInUpperCase = 'THE_TRADE_DESK';
 const advertiserId = 'test-advertiser-id';
 const dataProviderId = 'rudderstack';
 const segmentName = 'test-segment';
-const trackerId = 'test-trackerId';
 const sampleDestination = {
   Config: {
     advertiserId,
@@ -11,7 +10,6 @@ const sampleDestination = {
     dataServer: 'apac',
     ttlInDays: 30,
     audienceId: segmentName,
-    trackerId,
   },
   DestinationDefinition: { Config: { cdkV2Enabled: true } },
 };
@@ -35,73 +33,12 @@ const sampleContext = {
   sources: sampleSource,
 };
 
-const sampleContextForConversion = {
-  app: {
-    build: '1.0.0',
-    name: 'RudderLabs Android SDK',
-    namespace: 'com.rudderlabs.javascript',
-    version: '1.0.5',
-  },
-  device: {
-    adTrackingEnabled: true,
-    advertisingId: '3f034872-5e28-45a1-9eda-ce22a3e36d1a',
-    id: '3f034872-5e28-45a1-9eda-ce22a3e36d1a',
-    manufacturer: 'Google',
-    model: 'AOSP on IA Emulator',
-    name: 'generic_x86_arm',
-    type: 'ios',
-    attTrackingStatus: 3,
-  },
-  externalId: [
-    {
-      type: 'daid',
-      id: 'test-daid',
-    },
-  ],
-  ip: '0.0.0.0',
-  page: {
-    referrer: 'https://docs.rudderstack.com/destinations/trade_desk',
-  },
-  library: {
-    name: 'RudderLabs JavaScript SDK',
-    version: '1.0.5',
-  },
-  locale: 'en-GB',
-  os: {
-    name: '',
-    version: '',
-  },
-  screen: {
-    density: 2,
-  },
-  userAgent:
-    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.117 Safari/537.36',
-};
-
-const integrationObject = {
-  All: true,
-  THE_TRADE_DESK: {
-    policies: ['LDU'],
-    region: 'US-CA',
-    privacy_settings: [
-      {
-        privacy_type: 'GDPR',
-        is_applicable: 1,
-        consent_string: 'ok',
-      },
-    ],
-  },
-};
-
 export {
   destType,
   destTypeInUpperCase,
   advertiserId,
   dataProviderId,
   segmentName,
-  trackerId,
   sampleDestination,
   sampleContext,
-  sampleContextForConversion,
-  integrationObject,
 };
