+++ test/integrations/destinations/the_trade_desk/delivery/data.ts
@@ -5,7 +5,6 @@ import {
   dataProviderId,
   segmentName,
   sampleDestination,
-  trackerId,
 } from '../common';
 
 beforeAll(() => {
@@ -235,171 +234,6 @@ export const data = [
               destType: destTypeInUpperCase,
               destinationId: 'Non-determininable',
               errorCategory: 'platform',
-              feature: 'dataDelivery',
-              implementation: 'native',
-              module: 'destination',
-              workspaceId: 'Non-determininable',
-            },
-            status: 400,
-          },
-        },
-      },
-    },
-  },
-  {
-    name: destType,
-    description: 'Successful delivery of realtime conversion event to Trade Desk',
-    feature: 'dataDelivery',
-    module: 'destination',
-    version: 'v0',
-    input: {
-      request: {
-        body: {
-          version: '1',
-          type: 'REST',
-          method: 'POST',
-          endpoint: 'https://insight.adsrvr.org/track/realtimeconversion',
-          headers: {},
-          params: {},
-          body: {
-            JSON: {
-              data: [
-                {
-                  tracker_id: trackerId,
-                  adv: advertiserId,
-                  currency: 'USD',
-                  event_name: 'viewitem',
-                  value: 249.95000000000002,
-                  items: [
-                    {
-                      item_code: '622c6f5d5cf86a4c77358033',
-                      name: 'Cones of Dunshire',
-                      qty: 5,
-                      price: 49.99,
-                    },
-                  ],
-                  category: 'Games',
-                  brand: 'Wyatt Games',
-                  variant: 'exapansion pack',
-                  coupon: 'PREORDER15',
-                  position: 1,
-                  url: 'https://www.website.com/product/path',
-                  image_url: 'https://www.website.com/product/path.webp',
-                },
-              ],
-            },
-            JSON_ARRAY: {},
-            XML: {},
-            FORM: {},
-          },
-          files: {},
-        },
-        method: 'POST',
-      },
-    },
-    output: {
-      response: {
-        status: 200,
-        body: {
-          output: {
-            destinationResponse: {
-              response: {
-                Message: null,
-                EventResponses: [],
-              },
-              status: 200,
-            },
-            message: 'Request Processed Successfully',
-            status: 200,
-          },
-        },
-      },
-    },
-  },
-  {
-    name: destType,
-    description:
-      'Error response from the Trade Desk due to invalid real time conversion event payload',
-    feature: 'dataDelivery',
-    module: 'destination',
-    version: 'v0',
-    input: {
-      request: {
-        body: {
-          version: '1',
-          type: 'REST',
-          method: 'POST',
-          endpoint: 'https://insight.adsrvr.org/track/realtimeconversion',
-          headers: {},
-          params: {},
-          body: {
-            JSON: {
-              data: [
-                {
-                  tracker_id: trackerId,
-                  adv: advertiserId,
-                  currency: 'USD',
-                  event_name: 'viewitem',
-                  value: 249.95000000000002,
-                  items: [
-                    {
-                      item_code: '622c6f5d5cf86a4c77358033',
-                      name: 'Cones of Dunshire',
-                      qty: 5,
-                      price: 49.99,
-                    },
-                  ],
-                  category: 'Games',
-                  brand: 'Wyatt Games',
-                  variant: 'exapansion pack',
-                  coupon: 'PREORDER15',
-                  position: 1,
-                  url: 'https://www.website.com/product/path',
-                  image_url: 'https://www.website.com/product/path.webp',
-                  privacy_settings: [{}],
-                },
-              ],
-            },
-            JSON_ARRAY: {},
-            XML: {},
-            FORM: {},
-          },
-          files: {},
-        },
-        method: 'POST',
-      },
-    },
-    output: {
-      response: {
-        status: 400,
-        body: {
-          output: {
-            destinationResponse: {
-              response: {
-                Message: null,
-                EventResponses: [
-                  {
-                    EventIndex: 0,
-                    EventErrors: [
-                      {
-                        Error: 'InvalidPrivacySetting',
-                        ErrorMessage: 'The request has an invalid privacy setting.',
-                      },
-                    ],
-                    EventWarnings: [],
-                    Successful: false,
-                  },
-                ],
-              },
-              status: 400,
-            },
-            message:
-              'Request failed with status: 400 due to {"Message":null,"EventResponses":[{"EventIndex":0,"EventErrors":[{"Error":"InvalidPrivacySetting","ErrorMessage":"The request has an invalid privacy setting."}],"EventWarnings":[],"Successful":false}]}',
-            statTags: {
-              destType: destTypeInUpperCase,
-              destinationId: 'Non-determininable',
-              errorCategory: 'network',
-              errorType: 'aborted',
               feature: 'dataDelivery',
               implementation: 'native',
               module: 'destination',
