+++ test/integrations/destinations/the_trade_desk/network.ts
@@ -1,4 +1,4 @@
-import { destType, advertiserId, dataProviderId, segmentName, trackerId } from './common';
+import { destType, advertiserId, dataProviderId, segmentName } from './common';
 
 export const networkCallsData = [
   {
@@ -101,109 +101,6 @@ export const networkCallsData = [
       data: { FailedLines: [{ ErrorCode: 'MissingUserId', Message: 'Invalid UID2, item #2' }] },
       status: 200,
       statusText: 'Ok',
-    },
-  },
-  {
-    httpReq: {
-      url: 'https://insight.adsrvr.org/track/realtimeconversion',
-      data: {
-        data: [
-          {
-            tracker_id: trackerId,
-            adv: advertiserId,
-            currency: 'USD',
-            event_name: 'viewitem',
-            value: 249.95000000000002,
-            items: [
-              {
-                item_code: '622c6f5d5cf86a4c77358033',
-                name: 'Cones of Dunshire',
-                qty: 5,
-                price: 49.99,
-              },
-            ],
-            category: 'Games',
-            brand: 'Wyatt Games',
-            variant: 'exapansion pack',
-            coupon: 'PREORDER15',
-            position: 1,
-            url: 'https://www.website.com/product/path',
-            image_url: 'https://www.website.com/product/path.webp',
-          },
-        ],
-      },
-      params: { destination: destType },
-      headers: {
-        'Content-Type': 'application/json',
-        'User-Agent': 'RudderLabs',
-      },
-      method: 'POST',
-    },
-    httpRes: {
-      data: {
-        Message: null,
-        EventResponses: [],
-      },
-      status: 200,
-      statusText: 'OK',
-    },
-  },
-  {
-    httpReq: {
-      url: 'https://insight.adsrvr.org/track/realtimeconversion',
-      data: {
-        data: [
-          {
-            tracker_id: trackerId,
-            adv: advertiserId,
-            currency: 'USD',
-            event_name: 'viewitem',
-            value: 249.95000000000002,
-            items: [
-              {
-                item_code: '622c6f5d5cf86a4c77358033',
-                name: 'Cones of Dunshire',
-                qty: 5,
-                price: 49.99,
-              },
-            ],
-            category: 'Games',
-            brand: 'Wyatt Games',
-            variant: 'exapansion pack',
-            coupon: 'PREORDER15',
-            position: 1,
-            url: 'https://www.website.com/product/path',
-            image_url: 'https://www.website.com/product/path.webp',
-            privacy_settings: [{}],
-          },
-        ],
-      },
-      params: { destination: destType },
-      headers: {
-        'Content-Type': 'application/json',
-        'User-Agent': 'RudderLabs',
-      },
-      method: 'POST',
-    },
-    httpRes: {
-      data: {
-        Message: null,
-        EventResponses: [
-          {
-            EventIndex: 0,
-            EventErrors: [
-              {
-                Error: 'InvalidPrivacySetting',
-                ErrorMessage: 'The request has an invalid privacy setting.',
-              },
-            ],
-            EventWarnings: [],
-            Successful: false,
-          },
-        ],
-      },
-      status: 400,
-      statusText: 'Bad Request',
     },
   },
 ];
